Struct reroute::Router

pub struct Router {
    pub not_found: Option<NotFoundFn>,
    // some fields omitted
}

The Router struct contains the information for your app to route requests properly based on their HTTP method and matching route. It allows the use of a custom 404 handler if desired but provides a default as well.

Under the hood a Router uses a RegexSet to match URI's that come in to the instance of the hyper server. Because of this, it has the potential to match multiple patterns that you provide. It will call the first handler that it matches against so the order in which you add routes matters.

Fields

not_found

A custom 404 handler that you can provide.

Methods

impl Router

fn new() -> Router

Construct a new Router to maintain the routes and their handler functions.

fn add_route<H>(&mut self, verb: Method, route: &str, handler: H) where H: Fn(Request, Response, Captures) + Send + Sync + 'static

Add a route to the router and give it a function to call when the route is matched against. You can call this explicitly or use the convenience methods defined below.

fn options<H>(&mut self, route: &str, handler: H) where H: Fn(Request, Response, Captures) + Send + Sync + 'static

A convenience method for OPTIONS requests.

fn get<H>(&mut self, route: &str, handler: H) where H: Fn(Request, Response, Captures) + Send + Sync + 'static

A convenience method for GET requests.

fn post<H>(&mut self, route: &str, handler: H) where H: Fn(Request, Response, Captures) + Send + Sync + 'static

A convenience method for POST requests.

fn put<H>(&mut self, route: &str, handler: H) where H: Fn(Request, Response, Captures) + Send + Sync + 'static

A convenience method for PUT requests.

fn delete<H>(&mut self, route: &str, handler: H) where H: Fn(Request, Response, Captures) + Send + Sync + 'static

A convenience method for DELETE requests.

fn patch<H>(&mut self, route: &str, handler: H) where H: Fn(Request, Response, Captures) + Send + Sync + 'static

A convenience method for PATCH requests.

fn finalize(&mut self) -> Result<(), RouterError>

This function ensures that a valid RegexSet could be made from the route vector that was built while using the functions that add routes. It requires that there exist at least one route so that the RegexSet can be successfully constructed.

It will also ensure that there is a handler for routes that do not match any available in the set.

fn add_not_found(&mut self, not_found: NotFoundFn)

Add a function to handle routes that get no matches.